requirements.txt
flask
flask-swagger-ui
mysql-connector-python
